requirement.txt
amqp==2.6.1
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
bcrypt==3.1.7
cached-property==1.5.1
celery==4.4.7
cement==2.8.2
chardet==3.0.4
Django==3.0.4
django-cors-headers==3.2.1
django-postgrespool2==1.0.0
django-redis==4.11.0
djangorestframework==3.11.0
flake8==3.7.9
jsonschema==3.2.0
kafka-python==1.4.7
marshmallow==3.3.0
passlib==1.7.2
psycopg2-binary==2.8.5
pyflakes==2.1.1
redis==3.5.3
requests==2.23.0
urllib3==1.25.8
